server/requirements.txt
beautifulsoup4==4.9.3
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
google-api-core==1.23.0
google-auth==1.22.1
google-cloud-texttospeech==2.2.0
googleapis-common-protos==1.52.0
grpcio==1.33.1
idna==2.10
libcst==0.3.13
mypy-extensions==0.4.3
proto-plus==1.11.0
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydub==0.24.1
python-vlc==3.0.11115
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
rsa==4.6
simpleaudio==1.0.4
six==1.15.0
soupsieve==2.0.1
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.25.11
urlparser==0.1.2